Former-commit-id: 7cea5b3dffc38a283d302206d1181b7b3f9dbbc4

+++ Sources/Internal/Platform/Qt5/QtLayer.cpp
@@ -164,7 +164,7 @@ void QtLayer::KeyPressed(char16 key, int32 count, uint64 timestamp)
     UIEvent ev;
     ev.keyChar = 0;
     ev.phase = UIEvent::PHASE_KEYCHAR;
-    ev.timestamp = timestamp;
+    ev.timestamp = static_cast<float64>(timestamp);
     ev.tapCount = 1;
     ev.tid = key;
